requirements.txt
streamlit
pytesseract
pdf2image
spacy
opencv-python
google-generativeai